vapoursynth: split python module, mpv: use it
previously, vapoursynth was not a python module in how most python
libraries are in nixpkgs. it would only work as part of a library or
other program, and not with its included `vspipe` program.

this fixes it by adding a new output for its python module and uses
toPythonModule to allow it to work standalone, and without conflicting
with other python packages in a nix-shell